Update Documentation index
Imagine you looking to get into something, you dont have a lot of experience and want to read the docs.
Well, theres usually a wiki or a docs folder.
Unfortunally, this one is a bad joke, especially for newcomers. 

So in order to at least work around this, I present you a simple fix for a 15yo problem.