workaround macOS 10.14 performance bug
Since macOS 10.14 [NSView getRectsBeingDrawn:count:] just returns one combined rectangle.
This can be expensive, so we track the invalidated rectangles and check if the big combined rectangle is the same as the rectangles we tracked. Then if the area of our tracked rectangles is smaller then the area of the big combined one, we draw the individual rectangles.