fix cairo drawing error when another frame closes
If you have two CFrames open (for example two instances of the same
plugin), and close one of them, the drawing in the other one starts to
fail, with "the target surface has been finished" messages. The fix is
simple: just remove the call to cairo_device_finish.
cairo_device_destroy will call cairo_device_finish when it has decreased
the refcount to zero.