fix upside-down initial draw in FL Studio 2025
When using FL Studio 2025's "Detached" mode on macOS 14 and earlier,
on Intel or Rosetta only, the initial draw appeared upside down from
a user's point of view. Subsequent draws in response to mouse events
or parameters changing had the correct orientation.

It seems that setting geometryFlipped according to the parent layer's
contentsAreFlipped value does not guarantee correct drawing, because
the value of nsView.layer.contentsAreFlipped could be changed by the
host application or the OS. Looking for guidance from Apple, I found
WWDC 2012 Session 217 "Layer-Backed View" in which the presenter says
at timestamp 40:39 :-

> You don't have to manage geometryFlipped, and you probably shouldn't
> change it"

This patch implements what should be a more robust approach, which is
to leave geometryFlipped alone and instead query contentsAreFlipped
at draw time and adjust the CGContext's geometry if required.

I have verified that this works with DAWs that have been known to
cause upside down drawing:
* FL Studio 2025 (Intel) on macOS 14 and earlier
* Ableton Live 9 on macOS 10.14 and earlier
* Digital Performer 9 on macOS 10.14 and earlier
* NI Maschine